FINAL MAINPAGE v 1.02
This is the final result for the MainPage. It's basic, but we will improve it.

But, i cannot change or make any better the experience of the phones or tablets with @media.
In a future i will change it. But, until then, we have to conform with it. And also, the responsive is not done, but i have done some forms to the potential members and they say it's not necessary for now (if it's not necessary I'm not worried).